[docs] : src 폴더 내 types 폴더 생성을 통한 데이터, 컴포넌트 타입 관리 구조 변경
